+++ NEWS.md
@@ -1,3 +1,8 @@
+# bdftools 0.1.2   2021-07-27
+
+* Added `bdf_print_sample_compact()` to print using unicode block text
+  for more compact output
+
 # bdftools 0.1.1   2021-07-27
 
 * Characters used for 0/1 printing are now user-defined